Add set-playing command
I'm frequently annoyed that I'm unable to change where the playlist is
going to resume at after the queue is finished without interrupting
playback.

Running set-playing as part of win-add-q was tempting, but it'd be
confusing if it ran set-playing and win-add-Q didn't, and we also don't
want to run set-playing on win-add-Q if any tracks from the same
playlist/view are currently queued, and there currently isn't a way to
determine that.

I thought it might be nice to have the queue remember where queued items
are from and run set-playing when it's done playing the last track, but
I'm often happy with cmus's current behaviour where the queue doesn't
affect the play order of lib/pl at all, so it seems best to have
set-playing as just a standalone command. Another benefit of the
set-playing command as implemented in this commit is that one can also
change to other playlists or between lib/pl without interrupting
playback.